use dropzone-editor.ts
